feat: attachments (files/photos) and default new notes to short-term
- クリップアイコンからファイル・写真をSupabase Storageへアップロードし
  本文の下に一覧表示（画像はサムネイル、その他はファイル名+サイズ）
- 新規メモ作成時のデフォルトを長期保存から短期保存(7日)に変更

+++ components/AppShell.tsx
@@ -77,7 +77,7 @@ export default function AppShell({ userEmail }: { userEmail: string }) {
   async function handleCreate() {
     const folderId = view.type === "folder" ? view.folderId : null;
     const tag = view.type === "tag" ? [view.tag] : undefined;
-    const note = await createNote({ folder_id: folderId, type: "long", tags: tag });
+    const note = await createNote({ folder_id: folderId, type: "short", tags: tag });
     if (note) setSelectedId(note.id);
   }
